+++ main.py
@@ -1,5 +1,4 @@
-import tkinter as tk
-from gui import setup_gui
+from ui.gui import setup_gui
 
 if __name__ == "__main__":
-    setup_gui()
+    setup_gui()  # Inicializa a interface gráfica
